BAU: Add performance report resource
We need a better way to get performance platform data, I propose we
query the database, to do this we need an API endpoint which returns to
us the relevant data.

This creates a resource at `/v1/api/reports/performance-report` which
returns the volume of captured payments, the total amount in pence, and
the average amount in pence. e.g.:

```
{
  "total_volume": 1000,
  "total_amount": 1000,
  "average_amount": 1000
}
```

solo @tlwr