PP-3942 Introduce ServicePaymentReference object to connector
The new ServicePaymentReference is a simple immutable object that
wraps a String containing the reference specified by a service
when they create a payment.

Replace uses of Strings to represent the service payment reference
with the new ServicePaymentReference type to improve type-safety
and aid comprehension.

Introduce a ServicePaymentReferenceConverter to convert between
the new ServicePaymentReference type and Strings (and vice-versa)
when persisting/depersisting from the database.

Use Jackson’s ToStringSerializer to convert between the new
ServicePaymentReference type and Strings when serialising JSON.