PP-4054: Add lowercased gin indexes
Looking at production with the sql from a local postgresql
transaction we found it was having to do a filter. This was
due to it not having a trigram index for the lower cased version
of the reference or email.

So now adding an index specifically for the lowered version and
dropping the old unused ones.