PP-4072: See how many tests get fixed
This commit fixes CardAuthoriseResourceITest. How many other tests get fixed?
The build triggered by this commit will help us find out :)

The problem is we're setting the XRaySessionProfiler in the
ConnectorSessionCustomiser which starts a subsegment, does the database query,
then ends the subsegment. This fails if there was no segment started beforehand
- so we see in CardAuthoriseBaseService a `recorder.beginSegment` is happening
before the `preOperation` call which does database stuff, which is all good.
However we also needed a `recorder.beginSegment` before the `postOperation`
call as postOperation does database stuff too. Obviously having to do is very
bad - it would be better to have some aspect-oriented annotation around the Dao
classes that do the `beginSegment` and `endSegment`. This is still a work in
progress but I want to see how many tests get fixed by this commit as currently
we have `Tests run: 1015, Failures: 37, Errors: 142`.

@oswaldquek