PP-4183: Update capture process timings
Currently we only run the capture process every 150 to 200 seconds...
this means with 6 nodes and a batch size of 50 we'll only clear down 1.7
captures per second.

This reduces the average gap between runs to 15 and thus 20 tx per second.
We may need to run more or less than 6 nodes for 40 tx, but it is the right order
of magnitude.

It is important to have this running as it will give us more realistic performance.