PP-4248 Stop using internal JDK NotImplementedException
Make GuiceAppRule stop using sun.reflect.generics.reflectiveObjects.
NotImplementedException because it’s an JDK internal thing and not
available out of the box with Java 11.

The intention was probably to use the Apache Commons
NotImplementedException but Java’s own
UnsupportedOperationException is a better fit (we never plan to
implement those methods) so switch to that instead.