PP-4248 Upgrade EclipseLink from 2.6.4 to 2.7.3
Upgrade EclipseLink from 2.6.4 to 2.7.3 because 2.6.x versions do
not support Java 11.

Rather than bringing in all of EclipseLink, just depend on the JPA
and MOXy (XML) features because we don’t use the rest.

Rename the POM property jpa.version to eclipselink.version because
it represents the EclipseLink version, not the JPA version.

Update the versions of our Java commons libraries to a newer
release: the only change is that the utils submodule has had its
EclipseLink dependency upgraded from 2.6.4 to 2.7.3.

Nothing on https://www.eclipse.org/eclipselink/releases/2.7.php or
https://projects.eclipse.org/projects/ee4j.eclipselink/releases/2.7.3/review
indicates any backwards-compatibility issues that affect us: the
former page mentions that “Behavior of @XmlID and @XmlValue
annotations has been updated to be JAXB 2.2 RI compatible” but we
never use @XmlID and only use @XmlValue in a single test.

EclipseLink 2.7.3 supports JPA 2.2 (EclipseLink 2.6.4 supports JPA
2.1.1) but a perusal of the JPA documentation indicates no changes
to be concerned about.

Due to a signing issue, we now need to explicitly depend on JPA
2.1.1. See the comment in the POM file for the justification and
why this should cause any problems.