PP-4248 Make ChargeRefundServiceTest work with newer Mockito
When stubbing, newer Mockito releases seem to not consider null to
match any(Long.class). Some tests in ChargeRefundServiceTest stub
returning a RefundEntity from the (mocked) RefundDao by an ID that
turns out to actually be null, due to a RefundEntity only getting
an ID when the RefundDao persists it (the mock RefundDao does not
set an ID). Fix this by using a Mockito Answer to set the ID on
the RefundEntity (and in one test, on its spy double) to an
arbitrary (but non-null) value that can then be used in stubbing.
This is not great and ChargeRefundServiceTest could do with some
love but we want to minimise the number of changes we make at once.