PP-4276: Use direct debit connector's DropwizardJUnitRunner
This commit largely copies's dd-connector's DropwizardJUnitRunner which spins
up the dropwizard app and postgres container only once for all tests that use
it. This commit reduces the build time from ~20minutes to <10minutes.

In the ChargingITestBase we have had to make the
[worldpay|smartpay|epdq]MockClients static as some tests fail when all the
tests are being run. Probably some weird connection pooling issue.

@oswaldquek, @georgeracu