PP-4276 Code cleanup
- All classes that inherit from `ChargingITestBase` call `super.setup()`
when they have a method annotated with `@Before`
- The methods annotated with `@Before` have been renamed to `setup` so it
will override the method in the parent class. In some cases, the parent
method annotated with `@Before` was not called and some of the clients
were not instantiated, throwing NPE. JUnit uses Reflection for methods
annotated with `@Before` and when overriding it in a child class, the
parent functionality was not called. In some cases it was working because
it was called `setUp` or something else (mind the upper case `U`)
- Optimised imports and code style
- Removed locally instantiated client objects, as there's an `connectorRestApi`
object that gets created every time when a test is run. As we only use this
client objects to call the resources, it seems perfectly fine to use the
object that gets created anyway
- This change might help in the future in adding all the generic setup
required in the base class and test behaviour in the child classes that
also have the tests