Proposal for new validation pattern in connector
This commit illustrates how to use Dropwizard's built in validation
functionality to validate incoming requests. It is quite powerful, and
can do most of (probably all) the things we require of it.

In this commit I have taken one example endpoint
`/v1/frontend/charges/{chargeId}/status`, and refactored
validation for the PUT method.

I do not necessarily think that the validation logic is right
(in particular the constraint on the value of `new_status`),
but I have not changed it at this stage, just the implementation.

In order to change as little as possible, I have overridden the
dropwizard default exception mapper (see https://www.dropwizard.io/0.9.2/docs/manual/validation.html)
, so as  to be able to add my own `ConstraintViolationExceptionMapper`,
which allows me to map ConstraintViolationExceptions to 400 status
code. The returned message is slightly different from before, but
is not any worse IMHO.

In tandem with this change to the validation, I have changed how it
is tested. Instead of spinning up a whole application and db, I have
used the ResourceTestRule that provides a testable resource which can be
interacted with as if it is an http server, whilst allowing mocking
of it's dependencies. This allows tests to complete in ~400 ms as opposed to
~16 seconds before.

To do this I again had to do some fiddling with the exception mapper config,
(see https://www.dropwizard.io/1.3.5/docs/manual/testing.html)
which is the reason for the brillaintly named
`ResourceTestRuleWithCustomExceptionMappersBuilder`, which encapsulates the needed
changes.