PP-4276: Create and teardown databases at class level, not method level
To make this work, test data needs to have unique ids. Obviously, creating and
tearing down the database in the @BeforeClass / @AfterClass rather than @Before
/ @After methods makes the tests go faster as they are run only once per test
class, not test method.

Although test data ids should be unique, the original driver of this is was
that tests were failing because of JPA caching. To illustrate:

1. Test 1 creates the database
2. A charge with charge_id 123 is inserted directly via JDBI
3. The test retrieves the charge via JPA, which caches the charge with charge_id
4. Test 1 destroys the database
5. Test 2 creates the database
6. A charge with charge_id 123 is inserted directly via JDBI
7. The test retrieves the charge via JPA, which is the cached one from Test 1
8. Whatever we're testing fails because it's not the correct charge

@oswaldquek @georgeracu