Pull encrypted secrets from S3
This is a work in progress to test the possibility of storing larger
secrets encrypted in S3.

The secrets are encrypted in S3 with a symmetric key, the access to the
S3 is restricted but it is not considered the control point for access
to the secret.

The control point for access of the secret is access to the decryption
key.

The decryption key is stored in parameter store. So the secret has the
same control point as existing secrets stored in parameter store.