PP-4341 Make StripeGatewayClient more generic
The StripeGatewayClient is intended to eventually become the GatewayClient
for all gateways. For this to work, it needs to only now about 'clienty' things
- urls, payloads, headers... So this commit refactors the StripeGatewayClient
so that it only takes as arguments things that make sense for a generic client
to know about. The only mildly controversial one is the metrics prefix, which
we use within the client to emit correctly namespaced metrics. I think this is
fine.

This change in the client has prompted a small refactoring of StripePaymentProvider
- it now catches a checked exception from the client before deserialising
StripeErrorResponse - I think this is the right place for this to happen.

I have also taken the opportunity to extract out smaller methods to make the main
authorise method easier to read.