bau: Log number of active threads
The number of threads is currently set to threadsPerCpu (100) multiplied by
availableProcessors. In order to determine if we can get rid of
CardExecutorService, it is probably a good idea to find out if we ever hit
anything close to 100*availableProcessors. If we never we could probably delete
CardExecutorService as every incoming request (or thread) to
/v1/frontend/charges/{chargeId}/cards is served by a single instance of
CardResource and CardAuthoriseService and my thinking is that the
piping those requests into a singleton instance of CardExecutorService isn't
required.

@oswaldquek