PP-4341: Handle 4xx and 5xx errors from Stripe separately
We were already throwing a GatewayClientException from the StripeGatewayClient
to force the caller to handle errors. We managed to test 402 responses with the
help of https://stripe.com/docs/testing#cards-responses. 5xx are untestable
however and we are not sure what kind of response body is returned. We know how
to deserialize the json response from a 4xx error. We're not so sure about 5xx,
so we're just logging the entire response body using
`response.readEntity(String.class)`. To this end  we decided to map 4xx and 5xx
errors to GatewayClientException and DownstreamException respectively.

Also added javadoc to differentiate between the errors.

Needed to modify Stripe capture functionality as a result of these changes.

@oswaldquek @georgeracu