BAU - Remove logic from `ChargeEntity`
- I removed the two methods `hasStatus` from the `ChargeEntity` as this feels wrong.
The entity already has a `getStatus()` method which returns the current status, while
the two methods were used to actually check that the current status is in a list of
already known states. This can be easily accomplished by just using `contains` of
Collections framework
- Added a private constructor on `CancelServiceFunctions` to prevent people from
instantiating it as this class has only static methods
- Updated relevant tests to test on the actual String value that gets returned by the
`getStatus()` method, instead trying to compare a String (returned by the entity) with
a ChargeStatus value. Assert on the value you get and not on some other object.
- Reordered modifiers to comply with JLS. Will make it easier for future reading and
understanding of the code.