PP-4399 - Remove unnecessary object return
- When cancelling a charge, both resources are using a `ChargeCancelService`
and they both handle a response of type Optional<GatewayResponse<BaseCancelResponse>>

There are a few things that can be improved:
- This response type is returned even if the charge is not canceled with the
payment gateway, as it needs to satisfy the return signature. This has been
replaced with void now
- This object is not used by the resource. No matter what the result of the
cancellation, the resource will return `NoContent`. In case of an error, it
will log it, but will still return `NoContent. Now the logging is done in
the service, as there's no point to be logged a level up in the resource.
The resource returns `NoContent` and any exceptions like `ChargeNotFoundRuntimeException`
are thrown in the service and handled by the global exception handler
- Replaced `CancelServiceFunctions#changeStatusTo` in
`ChargeCancelService#nonGatewayCancel`
- Added @Transactional annotation to let Guice handle detached entities
and persist any changes through JPA. This replaces the transactional
behaviour of the replaced function