PP-4439 Refactored payment provider response
- At the moment we propagate `GatewayResponse<BaseAuthoriseResponse>` all
the way up to the resource. This is not a desired behaviour as this object
has a pretty convoluted structure, also issues compilation warnings and most
importantly, a service or a resource don't need to know about the response
from the payment provider (gateway). Also we were passing quite a few optional
method arguments around, which brings a bit of code smell. Here I extracted
into an object what a service needs to know from the payment gateway in order
to update the charge. This object will be converted into an even lighter one
that gets sent to the resource where the resource will deal with how to respond.
- In this craziness I also moved some specific methods from CardAuthoriseBaseService
into Card3dsResponseAuthService as it were specific to this class
- I extracted logic onto how to generate a `PaymentProviderAuthorisationResponse` into
this class as it will be into one place and anyone who needs one of this objects can
easily pass a gateway response into the factory method and voila, you a new object
- Simplified `AuthorisationResponse` to make use of the Java 8 Optional.ifPresent()
- Changed Apple related services to use the new object
- Inlined strings for logs and metrics
- There is also more work to be done to get rid of the generic object that gets
returned from the payment provider