PP-4548 Restore 3D Secure authorisation response logging
When logging the result of an attempt to authorise the result of a
3D Secure authentication with Worldpay, Smartpay or ePDQ, log more
information from the response itself or the error that meant we
could not get a response.

This restores the logging to a similar level of detail to what
we used to record (though for implementation reasons the logging is
not identical to previously).

Stripe never had this level of logging and this commit does not
attempt to change that (though we should examine this before we
go live with Stripe).

The logging relies on new stringification code in
Gateway3DSAuthorisationResponse. It would be nice if the new code
were similar to the code in the new CaptureResponse class that
helps with logging the results of attempts to capture, but the two
classes seem to have different ideas about how much they want to
know about the BaseResponse. We should attempt to bridge this
divide in the future.

Ideally there would be some tests for the logging (to stop it being
accidentally degraded again) but this commit does not add any in
the interests of restoring our logging detail quickly.