BAU: Add comment to capture logging noting its use by Sumo Logic
We now have a Sumo Logic saved search that depends on the logging
performed when we attempt to capture a payment. Add a comment
noting this so people think before tinkering with the logging and
breaking the saved search.

The comment is “// Used by Sumo Logic saved search”, which we’re
also using elsewhere in the code.