PP-4522: Add test and live token configuration
This commit allows the configuration of a test and live token. We want to be
able to support test and live accounts in production. The idea is that we can
do smoke test payments in production using a test token. At the time of this
commit, we aren't taking stripe payments in production yet. This commit only
concerns itself with adding the live token configuration without breaking
existing functionality. As such, all tokens used will be the test token.

@oswaldquek