PP-4550 Fix Stripe Client bug
- Fixed the case when `StripeGatewayClient` was hanging on requests after a while.
This behaviour has been linked to the Jersey Client where we didn't close
the `Response`.
- Given that we were propagating the `Response` object from the client a few
levels up to be handled by the caller of the client, the lifespan of this
object was long and it was keeping busy HTTP connections. The client now
returns a custom Response object, not a `javax.ws.rs.core.Response` one.
This new object type holds the response body, the status and family of the
status from the initial Response. The client will send back this new custom
object and any classes interested in using it will use the String payload
to map into the object type they need (if the payload is returned).
- Updated the tests to match the new response type
- Replaced a `WireMockRule` with a class rule, to keep the server running
between test classes. This change fixed a random test failure, that was
failing because the WireMock server failed to respond to the HTTP request
- Updated tests such that mocks will match against known URIs and not `any()`
- Deleted base test class as it was used only in one child class