PP-4573: Upgrade maven-surefire-plugin to 2.22.1
Upgrading to 3.0.0-M3 failed due to some maven version problem, so let's go to
the latest 2.x release of this plugin for now.

Future versions of failsafe deprecate the use of 'skipTests' for skipping
integration tests, so start using the new 'skipITs' property as well in
preparation for the release of v3.0.0 of the plugin.