PP-4553: Integration test for refund email notifications
SendRefundEmailTest tests that email notifications are sent for refunds. One
bad thing about this PR is that ConnectorApp and ConnectorModule have new
protected methods (`getModule` and `getNotifyClientFactory`) respectively. This
is code that enables us to mock out the NotifyClientFactory in the
SendRefundEmailTest. Unfortunately there is no nice way round doing this (it's
not Spring :( ) but it has been deemed essential to have an integration for
this functionality (mocking at unit test level is not enough). The other option
is to spin up a WireMock email notification server but unfortunately the notify
client needs to use https and enabling https in a wiremock and integration test
scenario is hugely not trivial.

@oswaldquek