BAU: Cleanups to gateway client timeout configuration code
Simplify the logic around setting metric names for the Jersey gateway clients.
We were falling back to "all" when there was no GatewayOperation specified, but
this was already an Optional and Optional can deal with fallbacks like this.

Replace logic branches with maps over Optionals. Hopefully this code is clearer
than what went before it.