BAU_clean_up_raw_generic_use
- `CardExecutorService` needn't be a generic class since the parameter is only
used within `CardExecutorService.execute()`. At present this class is being used
as a raw generic in mutiple places which presents compile and IDE
warnings. Also since the type isn't being specified the result of `.execute()`
is being cast which present a `unchecked cast warning`. To simplify the class
and its use, this makes `.excute()` a generic method and removes the generic
typing at the class level.