bau: functionalise things
A bit of refactoring "fun" (no pun intended).

This is a slightly controverisal commit and it's probably more for discussion
than me actually want this to be merged.

The idea of ChargeUtils is that it's "functional" - it's methods can be thought
of as [impure] functions that take some argument and return something (or not).
It does not hold state. Although currently, we pass in the databaseTestHelper
only for it use the other parameters passed in. In createNewRefund for example,
it's purpose is to generate a refund id which tests don't usually care about,
then save it. But currently pretty much all the paramters passed to it are not
being used by the method...they're being used by the databaseTestHelper
parameter being passed in!

This makes more code more "functional".

If people like it, maybe it could be merged!

@oswaldquek