PP-4512 Make AUTHORISATION 3DS REQUIRED expirable with gateway
In certain circumstances a charge that is in status `AUTHORISATION 3DS REQUIRED`
can exist in an authorised state with the gateway (specifically ePDQ - see
https://payments-platform.atlassian.net/browse/PP-4512).

If the charge is stuck in this state then the expiry sweep expires the charge in our
DB, but not with the gateway. This causes user pain, as the authorisation hold
is not released for days/weeks.

This commit adds `AUTHORISATION 3DS REQUIRED` as a 'gateway expirable' status.
In many cases this is likely to fail as the charge won't exist with the gateway
but the job seems to be robust to such failures (the failure is reported
in the HTTP response, but has no other effect).

I have added an explicit unit test for this case, partly to make it clear to
our future selves that this is deliberate, as it could be a bit confusing.