PP-4758: Change signature of GatewayClient
In an effort to make code simpler, we want to remove some functional idioms
that aren't particularly suited to Java. GatewayClient returns an
`Either<GatewayError, GatewayClient.Response>` which make upstream calling
classes complicated as we end up doing bi-Mapping and the like which is ugly
and confusing. When something goes wrong the java way is to throw exceptions
and catch them, which is what this commit does. GatewayClient now returns a
`GatewayClient.Response` which is a successful response. Unfortunately this is
a massive change as it's used by all the payment providers bar Stripe and there
is no way to reduce this.

With this in place we can deprecate the StripeGatewayClient which will be in
another PR.