PP-4512 Refactor charge expiry service to use ExpirableChargeStatus enum
Previously the source of truth for how a charge of a given status should
be expired was split between ChargeExpiryService and StatusFlow. I giess what I'm
trying to say is there was no source of truth.

Now, the source of truth is the enum ExpirableChargeStatus, from which one
can derive everything one might wish to know about the expirability of statuses.