PP-4512 Improve querying of ePDQ
Adds a method to the GatewayResponse builder so that queries
on payments that are in an error state are not themselves
interpreted as errors. This feels like a deeper problem, and I
have really only patched things for my needs since changing the
original method could have many unintended consequences.