PP-4560 Rename bank account and organisation details Stripe flags
In the Stripe setup JSON, rename "bank_account_details" to
"bank_account" and "organisation_vat_number_company_number" to
"organisation_details".

In the Stripe setup enum, rename BANK_ACCOUNT_DETAILS to
BANK_ACCOUNT and ORGANISATION_VAT_NUMBER_COMPANY_NUMBER to
ORGANISATION_DETAILS.

This is safe to do because nobody is using the JSON yet and the
none of the enum variants have been persisted to the database yet.

with @stephencdaly