PP-4791 Added custom BeanConstraintValidator
- Added annotation type definition for the custom constraint validator
for AuthCardDetails. This will work with annotation @Valid, instead of
calling the validator by hand every time we try to map AuthCardDetails
from external payload
- Constraint violations are now intercepted by javax and returned into
an array called `message` which is the default format. This requires a
small change in the way we assert in tests, as we need to assert that
the value is present in the array.
- Small code style changes: in-lined local variables, made default constructor
private to signal that this class should not be instantiated, removed
exception declaration from method signature